requirements.txt
pytest-html
python-dotenv
pytest-playwright
